Refactor ConferenceNewPage and add customize datepicker darkmode
