+++ src/components/ConferenceNewPage/types/Conference.ts
@@ -1,0 +1,16 @@
+export interface Conference {
+  name: string
+  url: string
+  city: string
+  country: string
+  online: boolean
+  startDate: Date | null | undefined
+  endDate: Date | null | undefined
+  topic: string
+  cfpUrl: string
+  cfpEndDate: Date | null | undefined
+  twitter: string
+  comment: string
+  cocUrl: string
+  offersSignLanguageOrCC: boolean
+}
